final_project/birds_telegram/requirements.txt
flask
gunicorn
Pillow
requests
tensorflow